% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoeikonal.R
\name{geogradient}
\alias{geogradient}
\title{Upwind Gradient of Distance on a Lon/Lat Grid}
\usage{
geogradient(dist, extent)
}
\arguments{
\item{dist}{A matrix of distances computed with \code{\link[=geoeikonal]{geoeikonal()}}.}

\item{extent}{A vector of grid extents}
}
\value{
A two element list containing the components of the gradient
}
\description{
Compute the (upwind) gradient of a solution to the Eikonal equation
computed with \code{\link[=geoeikonal]{geoeikonal()}}.
}
