% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoeikonal.R
\name{grdeikonal}
\alias{grdeikonal}
\title{Solve the Eikonal Equation on a Regular Grid}
\usage{
grdeikonal(cost, hx = 1, hy = 1)
}
\arguments{
\item{cost}{A matrix of travel costs - see details.}

\item{hx}{The grid spacing in the x direction}

\item{hy}{The grid spacing in the y direction}
}
\value{
A matrix of minimum travel costs to a source point
}
\description{
Solve the Eikonal Equation on a regular Grid.
}
\details{
This functions computes the minimum travel cost to each point in a
regular grid from a set of source points by solving the Eikonal
equation with a fast marching method.

The \code{cost} argument is a matrix of travel costs through the cells
of the grid, and is oriented so that x increases down the rows of
the grid, and y increases across the columns.  Positive entries of
the \code{cost} matrix give the cost of travel through the
corresponding cell of the grid, while \code{0} entries represent a
source point, and negative, \code{NA} or \code{Inf} entries represent
barriers that cannot be traversed.  The \code{hx} and \code{hy} arguments
give the grid spacings in the x and y directions.
}
\references{
Sethian, J.A. (1996), A fast marching level set method
for monotonically advancing fronts,
\emph{Proc. Natl. Acad. Sci.}  93 (4), 1591-1595.
}
