% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoeikonal.R
\name{grdgradient}
\alias{grdgradient}
\title{Upwind Gradient of Distance on a Lon/Lat Grid}
\usage{
grdgradient(dist, hx = 1, hy = 1)
}
\arguments{
\item{dist}{A matrix of distances computed with \code{\link[=grdeikonal]{grdeikonal()}}.}

\item{hx}{The grid spacing in the x direction}

\item{hy}{The grid spacing in the y direction}
}
\value{
A two element list containing the components of the gradient
}
\description{
Compute the (upwind) gradient of a solution to the Eikonal equation
computed with \code{\link[=grdeikonal]{grdeikonal()}}.
}
