% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoeikonal.R
\name{rsteikonal}
\alias{rsteikonal}
\title{Solve the Eikonal Equation on a Lon/Lat Raster}
\usage{
rsteikonal(cost)
}
\arguments{
\item{cost}{A \code{SpatRaster} with a longitude/latitude coordinate
reference system.}
}
\value{
A matrix of minimum travel costs to a source point
}
\description{
This functions computes the minimum travel cost to each point in a
longitude/latitude raster from a set of source points by solving
the Eikonal equation with a fast marching method.
}
\details{
The \code{cost} argument must be a \code{SpatRaster} with a
longitude/latitude coordinate reference system.  Positive entries
of the \code{cost} raster give the cost of travel through the
corresponding cell of the raster, while \code{0} entries represent a
source point, and negative, \code{NA} or \code{Inf} entries represent
barriers that cannot be traversed.

Distances are calculated assuming a spherical earth. If the
cost in each cell is the radius of the earth (6378.137km), then
the results are distances in kilometres.
}
