% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoeikonal.R
\name{rstgradient}
\alias{rstgradient}
\title{Upwind Gradient of Distance on a Lon/Lat Raster}
\usage{
rstgradient(dist)
}
\arguments{
\item{dist}{A matrix of distances computed with \code{\link[=rsteikonal]{rsteikonal()}}.}
}
\value{
A two element list containing the components of the gradient
}
\description{
Compute the (upwind) gradient of a solution to the Eikonal equation
computed with \code{\link[=rsteikonal]{rsteikonal()}}.
}
