% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoeikonal-package.R
\docType{package}
\name{geoeikonal-package}
\alias{geoeikonal-package}
\title{geoeikonal: Solve the Eikonal Equation on a Lon/Lat Grid}
\description{
Calculate shortest distance to target locations on an lon/lat grid by solving the eikonal equation with a fast marching algorithm.
}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

Authors:
\itemize{
  \item Dylan Wotherspoon \email{dylan.wotherspoon.au@gmail.com}
}

}
\keyword{internal}
